completed msghub client examples
