+++ cloud/sdr/sdr-data-ingest/example-go-clients/Makefile
